+++ packages/dds/merge-tree/src/test/Insertion.perf.spec.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { BenchmarkType, benchmark } from "@fluid-tools/benchmark";
+import { BenchmarkType, benchmarkDuration, benchmarkIt } from "@fluid-tools/benchmark";
 
 import { MergeTree } from "../mergeTree.js";
 import { MergeTreeDeltaType } from "../ops.js";
@@ -26,78 +26,95 @@ function constructTree(numOfSegments: number): MergeTree {
 const TREE_SIZE: number = 7500;
 
 describe("MergeTree insertion", () => {
-	benchmark({
+	benchmarkIt({
 		type: BenchmarkType.Measurement,
 		title: "insert into empty tree",
-		benchmarkFn: () => {
-			const emptyTree = new MergeTree();
-			emptyTree.insertSegments(
-				0,
-				[TextSegment.make("a")],
-				emptyTree.localPerspective,
-				{ seq: 0, clientId: 0 },
-				{ op: { type: MergeTreeDeltaType.INSERT } },
-			);
-		},
-	});
-
-	let startTree = constructTree(TREE_SIZE);
-	benchmark({
-		type: BenchmarkType.Measurement,
-		title: "insert at start of large tree",
-		benchmarkFn: () => {
-			for (let i = TREE_SIZE; i < TREE_SIZE + 25; i++) {
-				startTree.insertSegments(
+		...benchmarkDuration({
+			benchmarkFn: () => {
+				const emptyTree = new MergeTree();
+				emptyTree.insertSegments(
 					0,
 					[TextSegment.make("a")],
-					startTree.localPerspective,
-					{ seq: i + 1, clientId: 0 },
+					emptyTree.localPerspective,
+					{ seq: 0, clientId: 0 },
 					{ op: { type: MergeTreeDeltaType.INSERT } },
 				);
-			}
-		},
-		beforeEachBatch: () => {
-			startTree = constructTree(TREE_SIZE);
-		},
+			},
+		}),
 	});
 
-	let middleTree = constructTree(TREE_SIZE);
-	benchmark({
+	benchmarkIt({
+		type: BenchmarkType.Measurement,
+		title: "insert at start of large tree",
+		...benchmarkDuration({
+			benchmarkFnCustom: async (state) => {
+				let startTree = constructTree(TREE_SIZE);
+				let keepRunning: boolean;
+				do {
+					startTree = constructTree(TREE_SIZE);
+					keepRunning = state.timeBatch(() => {
+						for (let i = TREE_SIZE; i < TREE_SIZE + 25; i++) {
+							startTree.insertSegments(
+								0,
+								[TextSegment.make("a")],
+								startTree.localPerspective,
+								{ seq: i + 1, clientId: 0 },
+								{ op: { type: MergeTreeDeltaType.INSERT } },
+							);
+						}
+					});
+				} while (keepRunning);
+			},
+		}),
+	});
+
+	benchmarkIt({
 		type: BenchmarkType.Measurement,
 		title: "insert at middle of large tree",
-		benchmarkFn: () => {
-			for (let i = TREE_SIZE; i < TREE_SIZE + 25; i++) {
-				middleTree.insertSegments(
-					TREE_SIZE / 2,
-					[TextSegment.make("a")],
-					middleTree.localPerspective,
-					{ seq: i + 1, clientId: 0 },
-					{ op: { type: MergeTreeDeltaType.INSERT } },
-				);
-			}
-		},
-		beforeEachBatch: () => {
-			middleTree = constructTree(TREE_SIZE);
-		},
+		...benchmarkDuration({
+			benchmarkFnCustom: async (state) => {
+				let middleTree = constructTree(TREE_SIZE);
+				let keepRunning: boolean;
+				do {
+					middleTree = constructTree(TREE_SIZE);
+					keepRunning = state.timeBatch(() => {
+						for (let i = TREE_SIZE; i < TREE_SIZE + 25; i++) {
+							middleTree.insertSegments(
+								TREE_SIZE / 2,
+								[TextSegment.make("a")],
+								middleTree.localPerspective,
+								{ seq: i + 1, clientId: 0 },
+								{ op: { type: MergeTreeDeltaType.INSERT } },
+							);
+						}
+					});
+				} while (keepRunning);
+			},
+		}),
 	});
 
-	let endTree = constructTree(TREE_SIZE);
-	benchmark({
+	benchmarkIt({
 		type: BenchmarkType.Measurement,
 		title: "insert at end of large tree",
-		benchmarkFn: () => {
-			for (let i = TREE_SIZE; i < TREE_SIZE + 25; i++) {
-				endTree.insertSegments(
-					i,
-					[TextSegment.make("a")],
-					endTree.localPerspective,
-					{ seq: i + 1, clientId: 0 },
-					{ op: { type: MergeTreeDeltaType.INSERT } },
-				);
-			}
-		},
-		beforeEachBatch: () => {
-			endTree = constructTree(TREE_SIZE);
-		},
+		...benchmarkDuration({
+			benchmarkFnCustom: async (state) => {
+				let endTree = constructTree(TREE_SIZE);
+				let keepRunning: boolean;
+				do {
+					endTree = constructTree(TREE_SIZE);
+					keepRunning = state.timeBatch(() => {
+						for (let i = TREE_SIZE; i < TREE_SIZE + 25; i++) {
+							endTree.insertSegments(
+								i,
+								[TextSegment.make("a")],
+								endTree.localPerspective,
+								{ seq: i + 1, clientId: 0 },
+								{ op: { type: MergeTreeDeltaType.INSERT } },
+							);
+						}
+					});
+				} while (keepRunning);
+			},
+		}),
 	});
 });